move game file handling to a separate module
